move portable implementations of char utities out from gtest-port.h
